requirements.txt
python-telegram-bot==20.7
feedparser==6.0.11
aiohttp==3.9.1
